requirements.txt
streamlit
transformers
torch
Pillow
google-generativeai
